major documentation updates to all functions
