+++ .vscode/settings.json
@@ -6,6 +6,7 @@
         "Allon",
         "Alonso",
         "AlphaTauri",
+        "analysed",
         "Asini",
         "barh",
         "Boogity",
@@ -23,7 +24,9 @@
         "dashdot",
         "De Vries",
         "Dhabi",
+        "dimgrey",
         "Dinoco",
+        "Doimo",
         "dtype",
         "edgecolor",
         "Etihad",
@@ -60,10 +63,12 @@
         "Matty",
         "Mazespin",
         "McLaren",
+        "Minardi",
         "Motorrad",
         "ncol",
         "ncols",
         "Nevitt",
+        "Nico",
         "Nonw",
         "Norris",
         "nrows",
@@ -92,8 +97,10 @@
         "sprintking",
         "Spyker",
         "statisticslines",
+        "Stefanus",
         "Taberner",
         "Timothee",
+        "triplescore",
         "Tsunoda",
         "Verstappen",
         "weeklylineup",
@@ -110,6 +117,24 @@
         "здесь",
         "Мадлен",
         "Мазэспин",
-        "Попова"
+        "Попова",
+        "rosybrown",
+        "lightcoral",
+        "sandybrown",
+        "darkorange",
+        "darkolivegreen",
+        "palegreen",
+        "skyblue",
+        "dodgerblue",
+        "slategrey",
+        "royalblue",
+        "mediumblue",
+        "slateblue",
+        "blueviolet",
+        "hotpink",
+        "lawngreen",
+        "cadetblue",
+        "rebeccapurple",
+        "midnightblue"
     ]
 }
