--- Data/2024/Josh M/Test.json
@@ -1,14 +1,0 @@
-{
-    "Bahrain": {
-        "Driver 1": "",
-        "Driver 2": "",
-        "Driver 3": "",
-        "Driver 4": "",
-        "Driver 5": "",
-        "Constructor 1": "",
-        "Constructor 2": "",
-        "DRS Boost": "",
-        "Penalties": "",
-        "Perks": ""
-    }
-}
