--- Data/2024/Lineup/Results.json
@@ -1,70 +1,0 @@
-{
-    "Driver Points": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Values": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Team Points": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "RB": [],
-        "Red Bull": [],
-        "Stake": [],
-        "Williams": []
-    },
-    "Team Values": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "RB": [],
-        "Red Bull": [],
-        "Stake": [],
-        "Williams": []
-    }
-}
