--- Data/2024/Lineup/Statistics.json
@@ -1,206 +1,0 @@
-{
-    "Driver Points Per Value": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Sum Points": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Sum Values": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Average Points Per Value": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Average Points": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Driver Average Values": {
-        "Ocon": [],
-        "Gasly": [],
-        "Stroll": [],
-        "Alonso": [],
-        "Leclerc": [],
-        "Sainz": [],
-        "Magnussen": [],
-        "Hulkenberg": [],
-        "Norris": [],
-        "Piastri": [],
-        "Hamilton": [],
-        "Russell": [],
-        "Verstappen": [],
-        "Perez": [],
-        "Bottas": [],
-        "Guanyu": [],
-        "Tsunoda": [],
-        "Ricciardo": [],
-        "Albon": [],
-        "Sargeant": []
-    },
-    "Team Points Per Value": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    },
-    "Team Sum Points": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    },
-    "Team Sum Values": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    },
-    "Team Average Points Per Value": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    },
-    "Team Average Points": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    },
-    "Team Average Values": {
-        "Alpine": [],
-        "Aston Martin": [],
-        "Ferrari": [],
-        "Haas": [],
-        "McLaren": [],
-        "Mercedes": [],
-        "Red Bull": [],
-        "Stake": [],
-        "VCARB": [],
-        "Williams": []
-    }
-}
