--- Data/2024/Lineup_Weekly.json
@@ -1,39 +1,0 @@
-{
-    "Name": [
-        "Points",
-        "Value"
-    ],
-    "Race": [],
-    "Ocon": [],
-    "Gasly": [],
-    "Stroll": [],
-    "Alonso": [],
-    "Leclerc": [],
-    "Sainz": [],
-    "Magnussen": [],
-    "Hulkenberg": [],
-    "Norris": [],
-    "Piastri": [],
-    "Hamilton": [],
-    "Russell": [],
-    "Verstappen": [],
-    "Perez": [],
-    "Bottas": [],
-    "Guanyu": [],
-    "Tsunoda": [],
-    "Ricciardo": [],
-    "Albon": [],
-    "Sargeant": [],
-    "Alfa Romeo": [],
-    "AlphaTauri": [],
-    "Alpine": [],
-    "Aston Martin": [],
-    "Ferrari": [],
-    "Haas": [],
-    "McLaren": [],
-    "Mercedes": [],
-    "Red Bull": [],
-    "Stake": [],
-    "VCARB": [],
-    "Williams": []
-}
