+++ Finances/2021.json
@@ -1,0 +1,15 @@
+{
+    "Paid In": {"unknown": "unknown"},
+    "Winnings": {
+        "Stuart S": 28.58,
+        "Madlen P": 14.75,
+        "Josh H": 9.22,
+        "Will R": 25.35,
+        "James R": 7.83,
+        "Val M": 1.84,
+        "Jake M": 0.92,
+        "Nathan W": 0.92,
+        "Nick VL": 0.92,
+        "Josh M": 1.84
+    }
+}
