+++ Finances/2024.json
@@ -1,0 +1,29 @@
+{
+    "Paid In": {
+        "Madlen P": 5,
+        "Sam B": 1.22,
+        "James R": 15,
+        "Tom C": 5,
+        "Charlie L": 5,
+        "Andrew F": 5,
+        "Joe T": 5,
+        "Josh M": 5,
+        "Sebastian M": 5,
+        "Will R": 5,
+        "Patrick L": 5,
+        "Val M": 5,
+        "Jake M": 5,
+        "Phil M": 5,
+        "Chloe B": 5,
+        "Sam CJ": 5,
+        "Jo H": 5,
+        "Stefanus W": 5,
+        "Tim W": 5,
+        "Toby P": 5,
+        "Stuart S": 5,
+        "Cameron M":  15,
+        "Pete J": 5,
+        "Pierce C": 5
+    },
+    "Winnings": {}
+}
