--- Prizes.json
@@ -1,26 +1,0 @@
-{
-    "Maximum": ,
-    "Minimum": ,
-    "Second": ,
-    "Third": ,
-    "Average": ,
-}
-
-"
-league winner, second, third
-league loser, second loser,
-
-highest weekly
-lowest weekly
-
-highest value at abu dhabi
-highest average score across teams
-
-highest score sprint races
-highest score across certain races
-highest total score from top 5 in bahrain
-
-spot prizes at set races
-
-some will need to be functions, some will need races that change.
-"
