+++ config/Manager_Formats/Josh M.json
@@ -11,6 +11,9 @@
         "Oracle RedBull Ford-EcoPT",
         "Panasonic Toyota GRMN F1",
         "Becks Jaguar HSBC Racing",
-        "Test"
+        "Test",
+        "Petrobras Williams RBS F1",
+        "Siemens McLaren West",
+        "Vodafone Ferrari Marlboro"
     ]
 }
