+++ config/Manager_Formats/Will R.json
@@ -7,6 +7,8 @@
     "teams": [
         "Dinoco Piston Champions",
         "The Big One (TM)",
-        "Boogity Boogity Boogity"
+        "Boogity Boogity Boogity",
+        "Push To Pass",
+        "The Big One"
     ]
 }
