+++ lineup.py
@@ -10,6 +10,8 @@ from pathlib import Path
 def lineup_week(root : str,
                 year : str) -> None:
     """
+    Function Details
+    ================
     Calculate lineup results and plot figures.
 
     Use weekly lineup reports to calculate team and driver points and statistics
@@ -28,7 +30,7 @@ def lineup_week(root : str,
     --------
     update_weeklylineup
     update_lineup_stats
-    get_races_sofar
+    get_completed_races
     line_up_spreadsheet
     results_bar
     check_dir_exists
@@ -42,28 +44,50 @@ def lineup_week(root : str,
     -------
     None
 
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Update to documentation.
+
     """
-    results_dict = io.update_weeklylineup(
-        root_path=root,
-        year=year)
-    stats_dict = anal.update_lineup_stats(
-        root_path=root,
-        year=year,
-        results_dict=results_dict)
+
+    """ Load Config Files """
     info_path = Path(f'{root}/Info.json')
+    info_dict = (io.load_json(file_path=info_path))[f'{year}']
+    data_path = Path(f'{root}/Data/{year}')
     results_path = Path(f'{root}/Data/{year}/Lineup')
-    format_path = Path(f'{root}/Data/{year}/Lineup_Formats')
-    races_sofar, races = io.get_races_sofar(
-        file_path=info_path,
-        results_path=results_path)
+    format_path = Path(f'{root}/Config/Lineup_Formats')
+
+    """ Update the Weekly Lineup Points and Values """
+    results_dict = io.update_weeklylineup(
+        year=year,
+        info_dictionary=info_dict,
+        data_path=data_path,
+        results_path=results_path,
+        format_path=format_path)
+    stats_dict = anal.update_lineup_stats(
+        results_path=results_path,
+        results_dict=results_dict)
+
+    """ Check Completed Races """
+    completed_races = io.get_completed_races(
+        results_path=results_path,
+        info_dictionary=info_dict)
+
+    """ Create Spreadsheet """
     ss.line_up_spreadsheet(
         file_path=Path(f'{root}/Data/{year}/Lineup/Results.xlsx'),
         format_dir=format_path,
         races=races,
         results=results_dict,
         statistics=stats_dict)
-    for index, race in enumerate(races_sofar):
-        races = races_sofar[0: index + 1]
+
+    """ Plot Results Bar """
+    for index, race in enumerate(completed_races):
+        races = completed_races[0: index + 1]
         out_path = Path(f'{root}/Data/{year}/Figures/{race}')
         fp.check_dir_exists(dir_path=out_path)
         plot.results_bar(
@@ -72,8 +96,10 @@ def lineup_week(root : str,
             race=race,
             format_dir=format_path,
             out_path=out_path)
-    for index, race in enumerate(races_sofar):
-        races = races_sofar[0: index + 1]
+
+    """ Plot Statistics """
+    for index, race in enumerate(completed_races):
+        races = completed_races[0: index + 1]
         out_path = Path(f'{root}/Data/{year}/Figures/{race}')
         fp.check_dir_exists(dir_path=out_path)
         plot.lineupstats(
