+++ src/dataIO.py
@@ -519,7 +519,7 @@ def create_drivers_teams_weekly(lineup_path : str,
     return weekly_dictionary
 
 
-def updates_managers_weekly(dictionary_path : str,
+def updates_managers_weekly(team_dictionary : dict,
                             race_index : int,
                             races : list,
                             team_sheet : list) -> None:
@@ -531,8 +531,8 @@ def updates_managers_weekly(dictionary_path : str,
 
     Parameters
     ----------
-    dictionary_path : string
-        Path to manager team dictionary.
+    team_dictionary : dictionary
+        Manager team dictionary.
     race_index : int
         Race index in races list.
     races, team_sheet : list
@@ -563,8 +563,11 @@ def updates_managers_weekly(dictionary_path : str,
     Update History
     ==============
 
+    27/02/2024
+    ----------
+    Updated for manager dictionary path.
+
     """
-    team_dictionary = load_json(file_path=dictionary_path)
     index = race_index - 1
     while index >= 0:
         previous_race = races[index]
@@ -640,7 +643,7 @@ def creates_managers_weekly(race : str,
 
 def managers_weekly(info_dictionary : dict,
                     data_path : str,
-                    race_index : int):
+                    completed_races : list):
     """
     Function Details
     ================
@@ -652,9 +655,8 @@ def managers_weekly(info_dictionary : dict,
         Year information dictionary.
     data_path: string
         Path to yearly data folder.
-    race_index: int
-        Race index to update, note if 0 then the function creates a new blank
-        team sheet.
+    completed_races: list
+        List of races for which driver/team points/values already exist.
 
     Returns
     -------
@@ -687,25 +689,531 @@ def managers_weekly(info_dictionary : dict,
     so that it can determine whether to create a new blank dictionary or update
     an existing dictionary.
 
+    27/02/2024
+    ----------
+    Updated to check dictionary for completed races.
+
     """
     managers_dict = info_dictionary["Managers"]
     races = info_dictionary["Races"]
     for manager, teams in managers_dict.items():
         for team in teams:
             dictionary_path = Path(f'{data_path}/{manager}/{team}.json')
-            if dictionary_path.is_file():
-                updated_team_sheet = updates_managers_weekly(
-                    dictionary_path=dictionary_path,
-                    race_index=race_index,
-                    races=races,
-                    team_sheet=info_dictionary["Team"])
-                save_json_dicts(
-                    out_path=dictionary_path,
-                    dictionary=updated_team_sheet)
+            for index, race in enumerate(completed_races):
+                if dictionary_path.is_file():
+                    team_sheet = load_json(file_path=dictionary_path)
+                    if race in team_sheet.keys():
+                        pass
+                    else:
+                        updated_team_sheet = updates_managers_weekly(
+                            team_dictionary=team_sheet,
+                            race_index=index,
+                            races=races,
+                            team_sheet=info_dictionary["Team"])
+                        save_json_dicts(
+                            out_path=dictionary_path,
+                            dictionary=updated_team_sheet)                
+                else:
+                    blank_team_sheet = creates_managers_weekly(
+                        race=race,
+                        team_sheet=info_dictionary["Team"])
+                    save_json_dicts(
+                        out_path=dictionary_path,
+                        dictionary=blank_team_sheet)
+
+
+def get_completed_races(results_path : str,
+                        info_dictionary : dict) -> list:
+    """
+    Function Details
+    ================
+    Get a list of all completed races.
+
+    Uses the list of all races and the results file to determine which races
+    have been completed.
+
+    Parameters
+    ----------
+    results_path: string
+        Path to results directory for drivers and teams.
+    info_dictionary: dictionary
+        Season info dictionary.
+
+    Returns
+    -------
+    races_completed: list
+        List of completed races.
+
+    See Also
+    --------
+    None
+
+    Notes
+    -----
+    Uses the driver and team weekly results file to determine if a race has been
+    completed. If the file exists, the function assumes that the race has been
+    completed (or recorded) and will append the race name to a list of completed
+    races.
+
+    Example
+    -------
+    >>> races_so_far = get_completed_races(
+            results_path="/Path/To/Results/Directory",
+            info_dictionary=info_dict)
+    >>> races_so_far
+    ['Bahrain', 'China', 'Imola']
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    16/02/2024
+    ----------
+    Update to the function description for readability. Removed the load_json
+    info dictionary loading as this is not necessary and reduces RAM usage. Also
+    removed the returning of "races" as a variable, this is stored within the
+    info dictionary and does not need to be handled twice. This updated was
+    created by J.Male.
+
+    """
+    races = info_dictionary['Races']
+    races_completed = []
+    for race in races:
+        results = Path(f'{results_path}/{race}_Results.json')
+        if results.is_file():
+            races_completed.append(race)
+    return races_completed
+
+
+def manager_checked(statistics_dictionary : dict,
+                    data_path : str) -> None:
+    """
+    Function Details
+    ================
+    Create predicted manager scores.
+
+    Parameters
+    ----------
+    statistics_dictionary: dictionary
+        Managers statistics dictionary.
+    data_path: string
+        Path to data directory.
+    
+    Returns
+    -------
+    out_dict: dictionary
+        Predicted manager scores.
+    
+    See Also
+    --------
+    None
+
+    Notes
+    -----
+    None
+
+    Example
+    -------
+    None
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Update to documentation.
+
+    """
+    team_names = []
+    team_points = []
+    team_sum = statistics_dictionary["Team Sum Points"]
+    for manager, teams in team_sum.items():
+        for team, points in teams.items():
+            team_names.append(team)
+            team_points.append(points[-1])
+    zipped_lists = zip(team_points, team_names)
+    sorted_pairs = sorted(zipped_lists)
+    tuples = zip(*sorted_pairs)
+    sorted_points, sorted_names = [list(tuple) for tuple in tuples]
+    out_dict = {}
+    [
+        out_dict.update({n: p})
+        for n, p
+        in zip(sorted_names[::-1], sorted_points[::-1])]
+    save_json_dicts(
+        out_path=Path(f'{data_path}/Manager_Check.json'),
+        dictionary=out_dict)
+    return out_dict
+
+
+def update_weeklylineup(year : str,
+                        info_dictionary : dict,
+                        data_path : str,
+                        results_path : str,
+                        format_path : str) -> dict:
+    """
+    Function Details
+    ================
+    Update weekly driver and team points and values.
+
+    Uses json dictionaries to update the weekly results for all teams and
+    drivers. Uses the new weekly report and current results file to calculate
+    weekly values.
+
+    Parameters
+    ----------
+    year : string
+        Year of data collection.
+    info_dictionary: dictionary
+        Information dictionary for the year.
+    data_path, results_path, format_path: string
+        Path to data. Path to results directory. Path to format directory.
+
+    Returns
+    -------
+    results_dict : dictionary
+        Updated results dictionary containing weekly points and values.
+
+    See Also
+    --------
+    corrects_weekly
+    update_results_dict
+    load_json
+    save_json_dicts
+
+    Notes
+    -----
+    Reads in previous race results json files and ensure the results dictionary
+    is corrected, allowing for replacing values if mistakes are made. Then uses
+    the weekly lineup dictionary, if one exists, and uses the total values in
+    the results dictionary to correct the weekly lineup so that only points that
+    are scored in that race week are reported. This allows the owner to only
+    copy total points from the fantasy f1 website into the weekly lineup report.
+
+    Example
+    -------
+    >>> results_dictionary = update_weeklylineup(
+        root="/Path/To/Root/Directory",
+        year="example year")
+    >>> results_dictionary
+    {
+        "Driver 1": [1, 2, 56, 2],
+        "Team 2": [23, 19, 30, 45]
+    }
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Updated documentation.
+
+    """
+
+    """ Update Race Results """
+    results_dict = update_results_dict(
+        info_dictionary=info_dictionary,
+        results_path=results_path,
+        lineup_path=format_path)
+
+    """ Load New Week and Find Race """
+    weekly_lineup_dict = load_json(
+        file_path=Path(f'{data_path}/Lineup_Weekly.json'))
+    race = weekly_lineup_dict['Race']
+
+    """ If New Week """
+    if len(race) != 0:
+        print(race)
+        weekly_dict = corrects_weekly(
+            weekly_dictionary=weekly_lineup_dict,
+            info_dictionary=info_dictionary,
+            results_path=results_path)
+        save_json_dicts(
+            out_path=Path(f'{results_path}/{race[0]}_Results.json'),
+            dictionary=weekly_dict)
+        race_index = check_races(
+            race=race[0],
+            races=info_dictionary['Races'])
+        update_list = ['Driver', 'Team']
+        for category in update_list:
+            category_points = results_dict[f'{category} Points']
+            category_values = results_dict[f'{category} Values']
+            for key, inputs in weekly_dict.items():
+                if key == 'Name' or key == 'Race':
+                    pass
+                else:
+                    if key in category_points.keys():
+                        if len(category_points[key]) == race_index + 2:
+                            print(
+                                f'{key} {
+                                    (info_dictionary["Races"])[race_index]}'
+                                f'Points Recorded')
+                        else:
+                            category_points[key].append(inputs[0])
+                    if key in category_values.keys():
+                        if len(category_values[key]) == race_index + 2:
+                            print(
+                                f'{key} {
+                                    (info_dictionary["Races"])[race_index]}'
+                                f'Values Recorded')
+                        else:
+                            category_values[key].append(inputs[0])
+            results_dict.update({f'{category} Points': category_points})
+            results_dict.update({f'{category} Values': category_values})
+    else:
+        print('No Race To Report')
+
+    """ Save and Reset Weekly """
+    save_json_dicts(
+        out_path=Path(f'{results_path}/Results.json'),
+        dictionary=results_dict)
+    os.remove(path=Path(f'{data_path}/Lineup_Weekly.json'))
+    lineup_dictionary = create_drivers_teams_weekly(
+        lineup_path=format_path,
+        year=year)
+    save_json_dicts(
+        out_path=Path(f'{data_path}/Lineup_Weekly.json'),
+        dictionary=lineup_dictionary)
+    return results_dict
+
+
+def corrects_weekly(weekly_dictionary : dict,
+                    info_dictionary : dict,
+                    results_path : str) -> dict:
+    """
+    Function Details
+    ================
+    Corrects weekly lineup report.
+
+    Corrects weekly lineup report to only record points scored in that race week
+    into the results dictionary.
+
+    Parameters
+    ----------
+    weekly_dictionary, info_dictionary : dictionary
+        Weekly dictionary containing total points and current values. Info
+        dictionary containing season information, specifically list of races.
+    results_path : string
+        Path to results directory for lineup results dictionary.
+    
+    Returns
+    -------
+    individual_points_dict : dictionary
+        Corrected weekly points dictionary.
+    
+    See Also
+    --------
+    check_races
+    load_json
+    update_weeklylineup
+
+    Notes
+    -----
+    Updates the weekly lineup report containing total points and current values
+    for drivers and teams. Uses the results dictionary to calculate the points
+    scored on that particular race weekend by summing the total scored so far
+    and using the total points entered in the weekly report.
+
+    Example
+    -------
+    >>> weekly_lineup = corrects_weekly(
+        weekly_dictionary=lineup_weekly,
+        info_dictionary=info_dictionary,
+        results_path="/Path/To/Results/Directory")
+    >>> weekly_lineup
+    {
+        "Driver 1": [1, 2, 3, 4, 5],
+        "Team 1": [1, 2, 3, 4, 5]
+    }
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Updated Documentation.
+
+    """
+    race = weekly_dictionary['Race']
+    individual_points_dict = {}
+    race_index = check_races(
+        race=race[0],
+        races=info_dictionary['Races'])
+    if race_index == 0:
+        for key, inputs in weekly_dictionary.items():
+            individual_points_dict.update({key: inputs})
+    else:
+        previous_races = load_json(
+            file_path=Path(
+                f'{results_path}/Results.json'))
+        for key, inputs in weekly_dictionary.items():
+            if key == 'Name' or key == 'Race':
+                individual_points_dict.update({key: inputs})
             else:
-                blank_team_sheet = creates_managers_weekly(
-                    race=races[race_index],
-                    team_sheet=info_dictionary["Team"])
-                save_json_dicts(
-                    out_path=dictionary_path,
-                    dictionary=blank_team_sheet)
+                previous_results = dict(
+                    previous_races['Driver Points'],
+                    ** previous_races['Team Points'])
+                if inputs[0] == 'N/A':
+                    new_points = 0
+                    new_values = 0
+                else:
+                    new_points = inputs[0] - sum(previous_results[key])
+                    new_values = inputs[1]
+                individual_points_dict.update({key: [new_points, new_values]})
+    return individual_points_dict
+
+
+def check_races(race : str,
+                races : list) -> int:
+    """
+    Function Details
+    ================
+    Check the reported race.
+
+    Check reported race from list of races and find the race index.
+
+    Parameters
+    ----------
+    race : string
+        Race in the weekly report files or team entries.
+    races : list
+        List of all races in a season.
+    
+    Returns
+    -------
+    race_index : int
+        Race index from current season list.
+    
+    See Also
+    --------
+    get_races_sofar
+
+    Notes
+    -----
+    Uses list of all races in a season to return the list index of the current
+    race. Useful for adding or appending points/values lists.
+
+    Example
+    -------
+    >>> race = 'Bahrain'
+    >>> races = ['Bahrain', 'China', 'Imola']
+    >>> race_index = check_races(
+        race=race,
+        races=races)
+    >>> race_index
+    0
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Documentation updated.
+
+    """
+    race_index = []
+    for index, r in enumerate(races):
+        if r == race:
+            race_index.append(index)
+    return race_index[0]
+
+
+def update_results_dict(info_dictionary: dict,
+                        results_path: str,
+                        lineup_path: str) -> dict:
+    """
+    Function Details
+    ================
+    Update results dictionary from weekly reports.
+
+    Checks for all race reports and ensures results dictionary is up to date.
+
+    Parameters
+    ----------
+    info_dictionary: dictionary
+        Information dictionary for the season.
+    results_path, lineup_path : string
+        Paths to info dictionary, results directory, lineup format directory.
+    
+    Returns
+    -------
+    results_dict : dictionary
+        Results dictionary updated from weekly reports.
+    
+    See Also
+    --------
+    corrects_weekly
+    update_weeklylineup
+    load_json
+    get_completed_races
+    save_json_dicts
+
+    Notes
+    -----
+    Checks through the completed race reports in the results directory and makes
+    sure the results dictionary is correct from those. Allows the user to fix
+    incorrect or altered results without effort.
+
+    Example
+    -------
+    >>> results_dict = update_results_dict(
+        info_path="/Path/To/Info/Dictionary",
+        results_path="/Path/To/Results/Directory",
+        lineup_path="/Path/To/Lineup/Formats")
+    >>> results_dict
+    {
+        "Driver 1": [1, 2, 3, 4, 5],
+        "Team 1": [1, 2, 3, 4, 5]
+    }
+
+    ----------------------------------------------------------------------------
+    Update History
+    ==============
+
+    01/03/2024
+    ----------
+    Updated documentation and minor function name changes.
+
+    """
+
+    """ Find Config Files """
+    files = [
+        file for file in os.listdir(lineup_path) if 'Perks.json' not in file]
+    paths = [Path(f'{lineup_path}/{file}') for file in files]
+
+    """ Append Drivers and Teams """
+    teams = [os.path.splitext(os.path.basename(path))[0] for path in paths]
+    drivers = []
+    for team in teams:
+        file = load_json(file_path=Path(f'{lineup_path}/{team}.json'))
+        driver_names = file['drivers']
+        [drivers.append(driver) for driver in driver_names]
+
+    """ Get Completed Races """
+    completed_races = get_completed_races(
+        results_path=results_path,
+        info_dictionary=info_dictionary)
+
+    """ Update Race Results """
+    results_dict = load_json(file_path=Path(f'{results_path}/Results.json'))
+    for index, race in enumerate(completed_races):
+        race_results = load_json(
+            file_path=Path(f'{results_path}/{race}_Results.json'))
+        for key, values in race_results.items():
+            if key == 'Name' or key == 'Race':
+                pass
+            else:
+                if key in drivers:
+                    ((results_dict["Driver Points"])[key])[index] = values[0]
+                    ((results_dict["Driver Values"])[key])[index] = values[1]
+                if key in teams:
+                    ((results_dict["Team Points"])[key])[index] = values[0]
+                    ((results_dict["Team Values"])[key])[index] = values[1]
+    save_json_dicts(
+        out_path=Path(f'{results_path}/Results.json'),
+        dictionary=results_dict)
+    return results_dict
